Develop

+++ src/main/java/com/esoft/pcs/repository/PaddyPriceRepository.java
@@ -3,7 +3,9 @@ package com.esoft.pcs.repository;
 import com.esoft.pcs.models.PaddyPrice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaddyPriceRepository extends JpaRepository<PaddyPrice, Long>
+import java.util.List;
+
+public interface PaddyPriceRepository extends JpaRepository<PaddyPrice, Integer>
 {
     PaddyPrice getTopByDate(String date);
 }
